fix: Fix ObjC RUM operation options exposure
