+++ DatadogCore/Tests/Objc/ObjcAPITests/DDRUMMonitor+apiTests.m
@@ -6,7 +6,6 @@
 
 #import <XCTest/XCTest.h>
 @import DatadogRUM;
-@import DatadogInternal;
 
 @interface DDRUMMonitor_apiTests : XCTestCase
 @end
